video-replicator-standalone/backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
google-generativeai>=0.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
yt-dlp>=2023.0.0
requests>=2.28.0
python-multipart>=0.0.6
